Checks: Move highlighter to utils submodule
