+++ weblate/trans/templatetags/translations.py
@@ -35,7 +35,8 @@ from django.utils.translation import gettext, gettext_lazy, ngettext, pgettext
 from weblate.accounts.avatar import get_user_display
 from weblate.accounts.models import Profile
 from weblate.auth.models import User
-from weblate.checks import CHECKS, highlight_string
+from weblate.checks import CHECKS
+from weblate.checks.utils import highlight_string
 from weblate.trans.filter import get_filter_choice
 from weblate.trans.models import (
     Announcement,
